leagueforge/requirements.txt
Flask==3.0.0
gspread==5.12.0
google-auth==2.23.4

pandas==2.2.2
